tools/ttsServer/requirements.txt
# Coqui AI TTS (XTTS v2) requirements - using idiap fork
coqui-tts[codec]>=0.24.0
torch>=2.0.0
torchaudio>=2.0.0
soundfile>=0.12.1
pydub>=0.25.1
Flask>=2.2.5
numpy<2
